requirements.txt
python-telegram-bot[webhooks]>=20.0
openai>=1.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
pytest>=8.0
